move: useModal hook 파일 이동 (#9)
