+++ src/pages/form/my/index/_MyFormPage.tsx
@@ -1,6 +1,6 @@
 import { FormListLayout } from '@layout/FormListLayout';
 import { FormListItem } from '@api/form/types';
-import { useInfiniteScroll } from '@hooks/useInfiniteScroll';
+import { useInfiniteScroll } from '@/hooks/useInfiniteScroll';
 import { useGetInfiniteMyFormList } from '@query-hooks/form';
 
 export const MyFormPage = () => {
