+++ src/pages/news/index/_NewsPage.tsx
@@ -5,7 +5,7 @@ import { PostItem } from './PostItem';
 import { PostDetailResponse } from '@api/post/types';
 import { useSearchParams } from 'react-router-dom';
 import { isInCategory } from '@utils/isInCategory';
-import { useInfiniteScroll } from '@hooks/useInfiniteScroll';
+import { useInfiniteScroll } from '@/hooks/useInfiniteScroll';
 import { Intersection } from '@components/Intersection';
 import { Loader } from '@components/Loader';
