+++ src/pages/sign-up/email/MailEntry.tsx
@@ -6,7 +6,7 @@ import { usePostMail, usePostVerificationCode } from '@query-hooks/sign-up';
 import { useSignUpInfoStore } from '@stores/signUpInfo-stores';
 import { VerificationResponse } from '@api/sign-up/types';
 import { useNavigate } from 'react-router-dom';
-import { useModal } from '@hooks/useModal';
+import { useModal } from '@/hooks/modal/useModal';
 import { Loader } from '@components/Loader';
 
 export const MailEntry = ({
