+++ src/query-hooks/news-upload/index.ts
@@ -1,7 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
 import { ErrorResponse } from '@api/types';
-import { useModal } from '@hooks/useModal';
+import { useModal } from '@/hooks/modal/useModal';
 import { useNavigate } from 'react-router-dom';
 import { NewsUpload } from '@api/news-upload/types';
 import { News } from '@api/news-upload';
